backend/requirements.txt
Django==4.2.19
numpy==2.0.2
pandas==2.2.3
pip==25.0
python-dateutil==2.9.0.post0
pytz==2024.1
setuptools==72.1.0
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2023.3
wheel==0.45.1
djangorestframework==3.15.2
django-cors-headers==3.9